chore: setup monorepo
