+++ emlx/c_src/emlx_fast.cpp
