+++ emlx/lib/emlx/command_queue.ex
