+++ emlx/lib/emlx/fast.ex
