+++ emlx/test/emlx/to_binary_test.exs
